Moved get_user and is_logged_in to util.py, changed init import
